Adding short capillary length offset
Adding an offset (mm) for the short capillary tubes length. This is
needed as a correction of the tube length because the calculation of the
intersection of the tube with the tower's lateral planes is not perfect.
It does not intesect at the left or right point of the tube exactly. While
shortening the tube length also the tube placement in the tower is
changed to make sure the tube starts at the very end of the tower (back
face).