Adding a comment to bug in visualizer
Noticed that the root visualizer does not display tubes with
capillaryLength < 4.5 cm. This might lead to a misinterpretation of the
geometry. Added a comment to explain it in the code. I am going to add a
cut over the tube length for which tubes below it will not be placed.