Fix bug in C tube row placement
I used to stop all the tube placement (S+C) if there was not enough room
in y-direction to place a C tube on top of its S tube. This is a bug
because in some cases it left a single S tube in the last y-row at the
left side of the tower: it the C tube on top was not place also S tube
in x directin were not placed. Fixed by making the y-direction check
only applicable to the C tubes placement.