Change CMake usage
Change CMake usage from simple geometry example to a structure that
sources all the packages needed for geometry and simulation. Reorganize
code for geometry inside DRdetector/DRcalo directory. Keep
CMakeLists.txt file to build and compile only geometry with DD4hep (as
before) inside the onlygeocmake/ directory to be able to build and
visualize geometry locally (on lxplus I got a crash when trying to
visualize with geoDisplay).