Kill optical photon in C fiber clad
As I am counting optical photons internally relefcted in C fibers to
estimate the C signal it is important to not consider those created in
the cladding of C fiber. Otherwise the light yield estimation is wrong.
To do it I make cladding of C fiber sensible and kill optical photons in
it.