Use cp numbers in sensitive detector action
Using regex sensitive detector makes the methods volumeID and cellID
unusable. There is the possibility to get volIDs via brute force mapping
of G4Volume to TGeoVolume but it is slow. A long comment in the code
describes this problem. Therefore, I assigned copynumbers to TGeo
volumes in detector construction and use them a-la-Geant4 in the
sensitive detector action. Also removed an if statement to kill optical
photons in S fibers as they are not generated in S fibers.