Recreate VolID in SDAction using CpNumbers
The readout bit structure has been somplified removing tank and endcap
entries. In the geometry construction the copynumbers have been set
identical to the volIDs, or split into bits to allocated multiple volIDs
for the volume.
In the SDAction the copynumbers are split into bits and used to
recreated the volID with the BitFieldCoder. This slows down the
simulation by about 20%.